fix(hogql): type the flag-cache TTL setting as int
Every numeric instance setting is an int, and the settings API's type
cast handles int and bool but not float - a float setting edited from
the admin page would round-trip as a string. Whole seconds are all
this knob needs.

Generated-By: PostHog Desktop
Task-Id: 298f3504-9b69-4cb1-bb9d-908e0b4610d5

+++ posthog/hogql/database/database.py
@@ -357,7 +357,7 @@ def _cached_team_flag(flag_key: str, team: Team, evaluate: Callable[[], bool]) -
     # setting has its own short in-process cache, so this read costs no query per build.
     from posthog.models.instance_setting import get_instance_setting  # noqa: PLC0415
 
-    ttl = float(get_instance_setting("HOGQL_TEAM_FLAG_CACHE_TTL_SECONDS"))
+    ttl = int(get_instance_setting("HOGQL_TEAM_FLAG_CACHE_TTL_SECONDS"))
     if ttl <= 0:
         return evaluate()
     cache_key = (str(team.uuid), flag_key)
